biyo-v2/backend/requirements.txt
awesome-slugify==1.6.5
DistributedLock==1.2
Django==1.7.11
MySQL-python==1.2.5
Pillow==2.2.2
PyPDF2==1.20
PyYAML==3.11
TwitterAPI==2.3.6
argparse==1.2.1
astroid==1.4.3
beautifulsoup4==4.4.1
colorama==0.3.6
coverage==4.0.3
django-cors-headers==1.0.0
django-rest-auth==0.3.2
django-filter==0.10.0
django-annoying==0.8.5
django-bootstrap3==6.2.2
django-choices==1.1.12
django-coverage==1.2.4
django-debug-toolbar==1.4
django-extensions==1.6.1
django-jenkins==0.18.1
django-localflavor==1.2
django-m2m-history==0.3.2
django-oauth-tokens==0.6.2
django-picklefield==0.3.2
django-rest-swagger==0.3.4
django-taggit==0.17.6
django-twitter-api==0.2.8
djangorestframework==2.3.12
django-suit==0.2.16
flake8==2.1.0
foursquare==20130707
html5lib==0.999
httplib2==0.9.2
lazy-object-proxy==1.2.1
mccabe==0.2.1
oauthlib==1.0.3
pep8==1.7.0
pisa==3.0.33
poster==0.8.1
pyflakes==1.0.0
pylint==1.5.3
python-memcached==1.57
pytz==2015.7
reportlab==2.7
requests==2.6.0
requests-oauthlib==0.6.0
regex==2017.7.28
six==1.5.2
sqlparse==0.1.18
sorl-thumbnail==12.3
tweepy==3.5.0
wrapt==1.10.6
wsgiref==0.1.2
-e git+https://github.com/chrisglass/xhtml2pdf.git@4dc99c98a2276dbb3c9e8d34c44794756191afb0#egg=xhtml2pdf-dev
xlrd==0.9.4
yelpapi==1.3
Werkzeug==0.11.4
ipython==4.1.1
django-simple-history==1.8.0
django_countries==3.4.1
django-betterforms
